client/requirements.txt
GPUtil==1.4.0
GPUtil==1.4.0
numpy==2.1.3
psutil==6.1.0
py_cpuinfo==9.0.0
python-dotenv==1.0.1
Requests==2.32.3
server==0.0.0
torch==2.5.1
websockets==13.1
